\#12 Create battleship.main.Main.java to house the main method; delete main method from player
